requirements.txt
Django==1.10.2
djangorestframework==3.4.7
drf-nested-routers==0.11.1
django-allauth>=0.25.2,<0.26
django-annoying>=0.10.0
django-nested-admin==3.0.10
django-rest-auth==0.7.0
django-guardian==1.4.6
django-filter==1.0.1
Pillow==4.0.0